Opensbt/OPENSBT_ROOT/requirements.txt
deap==1.4.1
numpy==1.23.5
scipy==1.10.1
matplotlib==3.7.4
pandas==2.0.3
setuptools
scikit-learn==1.3.2
pymoo==0.6.0.1
pytest==8.0.0
psutil==5.9.8
docker

# surrogate
imblearn
tensorflow
keras-tuner
scikit-optimize
